+++ packages/contracts-rfq/test/FastBridgeV2.t.sol
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
 
+import {BridgeTransactionV2Lib} from "../contracts/libs/BridgeTransactionV2.sol";
+
 import {IFastBridge} from "../contracts/interfaces/IFastBridge.sol";
 
 // solhint-disable-next-line no-unused-import
@@ -44,11 +46,42 @@ abstract contract FastBridgeV2Test is Test, IFastBridgeV2Errors {
     IFastBridgeV2.BridgeParamsV2 internal tokenParamsV2;
     IFastBridgeV2.BridgeParamsV2 internal ethParamsV2;
 
+    bytes internal mockRequestV1;
+    bytes internal invalidRequestV2;
+    bytes internal mockRequestV3;
+
+    /// @notice We include an empty "test" function so that this contract does not appear in the coverage report.
+    function testFastBridgeV2Test() external {}
+
+    function createInvalidRequestV2(bytes memory requestV2) public pure returns (bytes memory result) {
+        // Copy everything but the last byte
+        result = new bytes(requestV2.length - 1);
+        for (uint256 i = 0; i < result.length; i++) {
+            result[i] = requestV2[i];
+        }
+    }
+
+    function createMockRequestV3(bytes memory requestV2) public pure returns (bytes memory result) {
+        result = new bytes(requestV2.length);
+        // Set the version to 3
+        result[0] = 0x00;
+        result[1] = 0x03;
+        // Copy the rest of the request
+        for (uint256 i = 2; i < result.length; i++) {
+            result[i] = requestV2[i];
+        }
+    }
+
     function setUp() public virtual {
         srcToken = new MockERC20("SrcToken", 6);
         dstToken = new MockERC20("DstToken", 6);
         createFixtures();
+        mockRequestV1 = abi.encode(extractV1(tokenTx));
+        // Invalid V2 request is formed before `createFixturesV2` to ensure it's not using callParams
+        invalidRequestV2 = createInvalidRequestV2(BridgeTransactionV2Lib.encodeV2(tokenTx));
         createFixturesV2();
+        // Mock V3 request is formed after `createFixturesV2` to ensure it's using callParams if needed
+        mockRequestV3 = createMockRequestV3(BridgeTransactionV2Lib.encodeV2(ethTx));
         fastBridge = deployFastBridge();
         configureFastBridge();
         mintTokens();
@@ -128,12 +161,14 @@ abstract contract FastBridgeV2Test is Test, IFastBridgeV2Errors {
             quoteRelayer: address(0),
             quoteExclusivitySeconds: 0,
             quoteId: bytes(""),
+            callValue: 0,
             callParams: bytes("")
         });
         ethParamsV2 = IFastBridgeV2.BridgeParamsV2({
             quoteRelayer: address(0),
             quoteExclusivitySeconds: 0,
             quoteId: bytes(""),
+            callValue: 0,
             callParams: bytes("")
         });
 
@@ -158,7 +193,6 @@ abstract contract FastBridgeV2Test is Test, IFastBridgeV2Errors {
         txV2.originAmount = txV1.originAmount;
         txV2.destAmount = txV1.destAmount;
         txV2.originFeeAmount = txV1.originFeeAmount;
-        txV2.sendChainGas = txV1.sendChainGas;
         txV2.deadline = txV1.deadline;
         txV2.nonce = txV1.nonce;
     }
@@ -166,6 +200,11 @@ abstract contract FastBridgeV2Test is Test, IFastBridgeV2Errors {
     function setTokenTestCallParams(bytes memory callParams) public {
         tokenParamsV2.callParams = callParams;
         tokenTx.callParams = callParams;
+    }
+
+    function setTokenTestCallValue(uint256 callValue) public {
+        tokenParamsV2.callValue = callValue;
+        tokenTx.callValue = callValue;
     }
 
     function setTokenTestExclusivityParams(address relayer, uint256 exclusivitySeconds) public {
@@ -180,6 +219,11 @@ abstract contract FastBridgeV2Test is Test, IFastBridgeV2Errors {
     function setEthTestCallParams(bytes memory callParams) public {
         ethParamsV2.callParams = callParams;
         ethTx.callParams = callParams;
+    }
+
+    function setEthTestCallValue(uint256 callValue) public {
+        ethParamsV2.callValue = callValue;
+        ethTx.callValue = callValue;
     }
 
     function setEthTestExclusivityParams(address relayer, uint256 exclusivitySeconds) public {
@@ -205,7 +249,6 @@ abstract contract FastBridgeV2Test is Test, IFastBridgeV2Errors {
         txV1.originAmount = txV2.originAmount;
         txV1.destAmount = txV2.destAmount;
         txV1.originFeeAmount = txV2.originFeeAmount;
-        txV1.sendChainGas = txV2.sendChainGas;
         txV1.deadline = txV2.deadline;
         txV1.nonce = txV2.nonce;
     }
@@ -221,11 +264,21 @@ abstract contract FastBridgeV2Test is Test, IFastBridgeV2Errors {
     }
 
     function getTxId(IFastBridgeV2.BridgeTransactionV2 memory bridgeTx) public pure returns (bytes32) {
-        return keccak256(abi.encode(bridgeTx));
+        return keccak256(BridgeTransactionV2Lib.encodeV2(bridgeTx));
     }
 
     function expectUnauthorized(address caller, bytes32 role) public {
         vm.expectRevert(abi.encodeWithSelector(IAccessControl.AccessControlUnauthorizedAccount.selector, caller, role));
+    }
+
+    function expectRevertInvalidEncodedTx() public {
+        vm.expectRevert(BridgeTransactionV2Lib.BridgeTransactionV2__InvalidEncodedTx.selector);
+    }
+
+    function expectRevertUnsupportedVersion(uint16 version) public {
+        vm.expectRevert(
+            abi.encodeWithSelector(BridgeTransactionV2Lib.BridgeTransactionV2__UnsupportedVersion.selector, version)
+        );
     }
 
     function cheatCollectedProtocolFees(address token, uint256 amount) public {