+++ packages/rest-api/src/routes/synapseTxIdRoute.ts
@@ -79,8 +79,6 @@ const router = express.Router()
  *               properties:
  *                 error:
  *                   type: string
- *                 details:
- *                   type: string
  */
 router.get(
   '/',